Massive refactor of TIM related files
